2.2.3-Mod语言文件 QF (#292)
* 戴森球→戴森云

戴森球（Dyson Sphere）和戴森云（Dyson Swarm）在技术上并非同种物件。
The variant closest to Dyson's original conception is the "Dyson swarm". It consists of a large number of independent constructs (usually solar power satellites and space habitats) orbiting in a dense formation around the star. This construction approach has advantages: components could be sized appropriately, and it can be constructed incrementally.

* 同步 工业屠宰场 QF

* Witchery 添加key

* 溢出停机 QF

* Revert "戴森球→戴森云"

This reverts commit 1e4815460c87c6e8f5146711c3877584715c4f62.